+++ .gitignore
@@ -110,6 +110,8 @@ ENV/
 env.bak/
 venv.bak/
 
+src/pyOpenAnnotate.egg-info
+
 # Spyder project settings
 .spyderproject
 .spyproject
