+++ example/Annotations/1.txt
@@ -1,0 +1,35 @@
+0 0.076 0.098 0.152 0.198
+0 0.074 0.086 0.15 0.172
+0 0.048 0.072 0.096 0.144
+0 0.068 0.064 0.136 0.128
+0 0.054 0.056 0.11 0.114
+0 0.048 0.086 0.096 0.174
+0 0.052 0.05 0.104 0.1
+0 0.064 0.06 0.128 0.122
+0 0.036 0.064 0.072 0.13
+0 0.062 0.082 0.126 0.164
+0 0.044 0.08 0.09 0.162
+0 0.056 0.052 0.112 0.104
+0 0.034 0.058 0.068 0.118
+0 0.028 0.062 0.056 0.126
+0 0.028 0.062 0.058 0.126
+0 0.042 0.05 0.084 0.1
+0 0.032 0.052 0.064 0.106
+0 0.032 0.074 0.066 0.148
+0 0.044 0.028 0.09 0.058
+0 0.036 0.042 0.072 0.084
+0 0.03 0.068 0.062 0.136
+0 0.018 0.058 0.036 0.118
+0 0.016 0.056 0.034 0.114
+0 0.032 0.022 0.066 0.046
+0 0.026 0.026 0.054 0.052
+0 0.028 0.032 0.058 0.064
+0 0.022 0.024 0.046 0.05
+0 0.024 0.024 0.048 0.05
+0 0.008 0.048 0.016 0.096
+0 0.018 0.018 0.038 0.036
+0 0.016 0.028 0.034 0.056
+0 0.004 0.01 0.01 0.022
+0 0.006 0.006 0.012 0.012
+0 0.006 0.004 0.014 0.01
+0 0.004 0.006 0.01 0.012
