+++ example/Annotations/10.txt
@@ -1,0 +1,26 @@
+0 0.046 0.124 0.094 0.248
+0 0.048 0.102 0.098 0.206
+0 0.102 0.048 0.206 0.096
+0 0.044 0.132 0.088 0.266
+0 0.064 0.122 0.128 0.246
+0 0.044 0.092 0.09 0.186
+0 0.06 0.106 0.12 0.214
+0 0.05 0.08 0.1 0.16
+0 0.054 0.06 0.11 0.122
+0 0.04 0.094 0.082 0.19
+0 0.042 0.088 0.086 0.176
+0 0.082 0.038 0.166 0.078
+0 0.068 0.042 0.138 0.086
+0 0.034 0.082 0.068 0.166
+0 0.038 0.072 0.076 0.144
+0 0.02 0.062 0.04 0.126
+0 0.012 0.032 0.024 0.066
+0 0.004 0.01 0.008 0.02
+0 0.008 0.006 0.016 0.012
+0 0.004 0.006 0.01 0.014
+0 0.004 0.006 0.01 0.012
+0 0.004 0.012 0.008 0.024
+0 0.006 0.008 0.012 0.016
+0 0.002 0.014 0.006 0.028
+0 0.004 0.006 0.01 0.012
+0 0.004 0.008 0.008 0.018
