+++ example/Annotations/2.txt
@@ -1,0 +1,28 @@
+0 0.146 0.064 0.292 0.128
+0 0.132 0.042 0.266 0.086
+0 0.092 0.056 0.184 0.114
+0 0.114 0.066 0.23 0.132
+0 0.094 0.052 0.188 0.106
+0 0.108 0.05 0.218 0.1
+0 0.054 0.056 0.11 0.114
+0 0.088 0.034 0.176 0.068
+0 0.096 0.028 0.192 0.058
+0 0.044 0.028 0.09 0.058
+0 0.036 0.042 0.072 0.084
+0 0.036 0.056 0.072 0.114
+0 0.078 0.018 0.156 0.038
+0 0.032 0.022 0.066 0.046
+0 0.022 0.024 0.046 0.05
+0 0.056 0.014 0.114 0.03
+0 0.018 0.018 0.038 0.036
+0 0.034 0.006 0.07 0.012
+0 0.004 0.014 0.01 0.03
+0 0.018 0.01 0.036 0.022
+0 0.004 0.01 0.008 0.022
+0 0.004 0.01 0.01 0.02
+0 0.002 0.01 0.006 0.02
+0 0.006 0.006 0.012 0.012
+0 0.006 0.004 0.014 0.01
+0 0.006 0.004 0.014 0.01
+0 0.006 0.006 0.012 0.014
+0 0.006 0.004 0.014 0.008
