+++ example/Annotations/3.txt
@@ -1,0 +1,48 @@
+0 0.074 0.102 0.15 0.206
+0 0.074 0.09 0.15 0.18
+0 0.066 0.066 0.134 0.132
+0 0.054 0.056 0.11 0.114
+0 0.046 0.092 0.094 0.184
+0 0.054 0.052 0.108 0.106
+0 0.064 0.066 0.128 0.134
+0 0.048 0.052 0.096 0.106
+0 0.05 0.052 0.102 0.106
+0 0.056 0.052 0.114 0.104
+0 0.062 0.086 0.126 0.172
+0 0.05 0.044 0.1 0.09
+0 0.056 0.054 0.112 0.108
+0 0.028 0.066 0.056 0.132
+0 0.032 0.062 0.066 0.124
+0 0.028 0.064 0.056 0.13
+0 0.05 0.052 0.1 0.106
+0 0.04 0.052 0.082 0.104
+0 0.032 0.054 0.064 0.11
+0 0.032 0.078 0.066 0.158
+0 0.044 0.046 0.088 0.092
+0 0.036 0.044 0.074 0.088
+0 0.044 0.028 0.09 0.058
+0 0.034 0.042 0.07 0.084
+0 0.03 0.072 0.06 0.144
+0 0.028 0.046 0.056 0.092
+0 0.018 0.062 0.036 0.126
+0 0.016 0.064 0.032 0.128
+0 0.024 0.036 0.05 0.074
+0 0.028 0.028 0.056 0.058
+0 0.028 0.034 0.056 0.07
+0 0.032 0.022 0.064 0.046
+0 0.012 0.066 0.024 0.134
+0 0.018 0.026 0.036 0.052
+0 0.006 0.044 0.014 0.09
+0 0.016 0.032 0.034 0.064
+0 0.01 0.028 0.022 0.056
+0 0.008 0.006 0.018 0.012
+0 0.004 0.01 0.01 0.022
+0 0.004 0.01 0.008 0.02
+0 0.004 0.01 0.008 0.02
+0 0.008 0.006 0.016 0.014
+0 0.004 0.008 0.008 0.018
+0 0.006 0.006 0.012 0.012
+0 0.008 0.014 0.018 0.028
+0 0.006 0.004 0.014 0.01
+0 0.004 0.006 0.008 0.014
+0 0.004 0.006 0.01 0.012
