+++ example/Annotations/4.txt
@@ -1,0 +1,44 @@
+0 0.06 0.152 0.12 0.306
+0 0.084 0.064 0.17 0.13
+0 0.078 0.07 0.156 0.14
+0 0.05 0.07 0.1 0.14
+0 0.042 0.106 0.084 0.212
+0 0.058 0.066 0.116 0.134
+0 0.05 0.064 0.102 0.13
+0 0.066 0.036 0.134 0.072
+0 0.04 0.068 0.082 0.136
+0 0.028 0.072 0.058 0.144
+0 0.054 0.044 0.11 0.09
+0 0.024 0.08 0.05 0.16
+0 0.046 0.044 0.092 0.088
+0 0.044 0.05 0.09 0.1
+0 0.046 0.042 0.092 0.084
+0 0.02 0.072 0.04 0.146
+0 0.032 0.058 0.066 0.116
+0 0.022 0.078 0.046 0.156
+0 0.048 0.028 0.098 0.056
+0 0.038 0.03 0.078 0.06
+0 0.028 0.06 0.058 0.12
+0 0.018 0.064 0.036 0.128
+0 0.02 0.052 0.04 0.104
+0 0.036 0.04 0.072 0.08
+0 0.018 0.082 0.036 0.164
+0 0.024 0.044 0.05 0.088
+0 0.032 0.036 0.066 0.074
+0 0.012 0.05 0.026 0.1
+0 0.016 0.036 0.032 0.072
+0 0.026 0.024 0.052 0.05
+0 0.022 0.026 0.046 0.054
+0 0.012 0.03 0.024 0.06
+0 0.012 0.02 0.026 0.042
+0 0.008 0.026 0.018 0.052
+0 0.012 0.018 0.024 0.038
+0 0.014 0.012 0.03 0.026
+0 0.006 0.01 0.014 0.022
+0 0.006 0.006 0.012 0.014
+0 0.004 0.006 0.008 0.014
+0 0.006 0.006 0.012 0.014
+0 0.006 0.006 0.014 0.014
+0 0.006 0.004 0.014 0.01
+0 0.004 0.006 0.01 0.012
+0 0.014 0.002 0.028 0.004
