+++ example/Annotations/5.txt
@@ -1,0 +1,15 @@
+0 0.128 0.056 0.258 0.112
+0 0.048 0.104 0.096 0.208
+0 0.138 0.072 0.276 0.144
+0 0.114 0.068 0.23 0.136
+0 0.086 0.05 0.174 0.102
+0 0.112 0.05 0.224 0.102
+0 0.088 0.036 0.176 0.072
+0 0.092 0.028 0.186 0.058
+0 0.036 0.108 0.072 0.216
+0 0.036 0.058 0.072 0.116
+0 0.078 0.018 0.156 0.038
+0 0.058 0.014 0.118 0.028
+0 0.042 0.006 0.084 0.012
+0 0.006 0.004 0.014 0.01
+0 0.002 0.01 0.006 0.02
