+++ example/Annotations/6.txt
@@ -1,0 +1,16 @@
+0 0.098 0.098 0.198 0.198
+0 0.088 0.064 0.176 0.128
+0 0.076 0.082 0.152 0.164
+0 0.064 0.064 0.128 0.13
+0 0.082 0.074 0.166 0.15
+0 0.074 0.09 0.148 0.18
+0 0.068 0.054 0.136 0.11
+0 0.056 0.03 0.114 0.062
+0 0.024 0.11 0.048 0.222
+0 0.048 0.032 0.098 0.064
+0 0.034 0.076 0.068 0.154
+0 0.008 0.028 0.016 0.056
+0 0.022 0.012 0.044 0.026
+0 0.02 0.006 0.042 0.014
+0 0.014 0.004 0.028 0.008
+0 0.006 0.004 0.014 0.008
