+++ example/Annotations/7.txt
@@ -1,0 +1,16 @@
+0 0.1 0.11 0.2 0.22
+0 0.112 0.058 0.226 0.118
+0 0.08 0.092 0.162 0.184
+0 0.12 0.046 0.242 0.094
+0 0.094 0.056 0.188 0.112
+0 0.104 0.044 0.208 0.088
+0 0.098 0.04 0.196 0.082
+0 0.1 0.032 0.202 0.066
+0 0.078 0.036 0.158 0.072
+0 0.102 0.068 0.206 0.136
+0 0.078 0.034 0.158 0.07
+0 0.084 0.042 0.17 0.086
+0 0.01 0.04 0.022 0.082
+0 0.014 0.012 0.028 0.026
+0 0.008 0.008 0.018 0.016
+0 0.008 0.01 0.016 0.02
