+++ example/Annotations/8.txt
@@ -1,0 +1,26 @@
+0 0.074 0.124 0.148 0.248
+0 0.064 0.056 0.128 0.114
+0 0.086 0.056 0.172 0.114
+0 0.082 0.066 0.166 0.132
+0 0.066 0.108 0.134 0.216
+0 0.072 0.074 0.144 0.148
+0 0.08 0.054 0.162 0.108
+0 0.062 0.044 0.124 0.088
+0 0.026 0.072 0.054 0.146
+0 0.044 0.054 0.09 0.11
+0 0.054 0.028 0.108 0.056
+0 0.058 0.058 0.116 0.118
+0 0.024 0.046 0.05 0.094
+0 0.054 0.054 0.11 0.11
+0 0.046 0.026 0.092 0.052
+0 0.034 0.026 0.07 0.052
+0 0.022 0.056 0.046 0.114
+0 0.008 0.026 0.016 0.052
+0 0.008 0.03 0.018 0.06
+0 0.016 0.018 0.034 0.036
+0 0.022 0.008 0.046 0.018
+0 0.004 0.026 0.01 0.054
+0 0.01 0.006 0.022 0.012
+0 0.006 0.006 0.014 0.014
+0 0.008 0.004 0.018 0.008
+0 0.004 0.006 0.01 0.012
