+++ example/Annotations/9.txt
@@ -1,0 +1,13 @@
+0 0.048 0.15 0.098 0.302
+0 0.066 0.094 0.132 0.19
+0 0.088 0.038 0.176 0.078
+0 0.082 0.036 0.164 0.074
+0 0.098 0.05 0.198 0.102
+0 0.08 0.034 0.16 0.068
+0 0.082 0.034 0.164 0.07
+0 0.06 0.044 0.122 0.09
+0 0.056 0.032 0.114 0.066
+0 0.104 0.056 0.208 0.112
+0 0.006 0.014 0.012 0.028
+0 0.01 0.01 0.022 0.022
+0 0.006 0.004 0.014 0.008
