+++ src/pyproject.toml
@@ -1,0 +1,28 @@
+[build-system]
+requires      = ["setuptools>=61.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "pyOpenAnnotate"
+version = "0.3.2"
+description = "Single class automated annotation tool using OpenCV"
+readme = "README.md"
+authors = [{ name = "Kukil Kashyap Borgohain", email = "kukilp213@gmail.com" }]
+license = { file = "LICENSE" }
+classifiers = [
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3",
+]
+keywords = ["Annotation", "Automated Annotation", "OpenCV Annotate"]
+dependencies = [
+    "opencv-python",
+    "natsort",
+]
+requires-python = ">=3.5"
+
+[project.urls]
+Homepage = "https://github.com/kXborg/pyOpenAnnotate"
+
+[project.scripts]
+annotate = "annotation.annotate:main"
